remove in_module mechanism
The procedure in_module was a central concept to the "new module system"
of 2010-03-19 that was to have made modules possible within the framework
of Conkeror's 'load' and 'require' mechanism.  However, there was a fatal
flaw in the design, and it really doesn't work after all.  The flaw was
that it didn't take account of hoisting, the property of javascript
whereby evaluation happens in two stages: first variables and functions
are defined, "hoisted" to the top of the scope, and then the code is
executed.  The in_module mechanism depended upon exceptions to stop the
load of a javascript file if it was being loaded into the wrong scope, but
because of hoisting, the crucial exception can only take place after all
of the variables and functions have been defined, thus the in_module idea
is not really a workable idea in javascript.

This patch removes the in_module mechanism.  A patch to follow will
introduce a new module system.  This patch leaves the module
casual-spelling broken, to be resolved by the next patch with the
replacement module system.

A related, but little-used feature, skip_module_load, was broken for the
same reason, and also removed.