tab-bar, new-tabs: set ordinal after adding new tab to tab bar
This works around a bug in XULRunner 1.9.x whereby ordinals fail to be
applied in some cases.

A related, but slightly different issue is here:

  https://bugzilla.mozilla.org/show_bug.cgi?id=555987

It seems likely that the fix for the above bugzilla bug also fixed the bug
we are seeing with tab-bar and new-tabs, because it only happens in
XULRunner < 2.0.