page-mode 'test' may now be an array
This provides simpler means for users to change the coverage of a
page-mode.  Instead of having to write a test function with all the
boilerplate involved in making sure the old test function gets called
properly, the 'test' property of a page-mode object is an array that can
be pushed and popped for customization.