Info.plist: set CFBundleShortVersionString to 1.0
According to the docs, this key is supposed to be a release version number
for the app, while CFBundleVersion may be an unreleased version number.
Apple's documentation is fairly specific about the format of a version
string (x.y.z) but we can't change our whole version system just for them.