hints-minibuffer-annotation-mode: inline hints URL display
During a hints interaction, conkeror can display the URL and other
information about the currently selected hint.  The information is now
displayed in the minibuffer, using minibuffer-annotation-mode.

This display replaces hints_url_panel; the content is the same, but the
new display looks better and it avoids obscuring part of the content
buffer, fixing http://bugs.conkeror.org/issue343.

To enable:

  hints_minibuffer_annotation_mode(true);

This replaces the former:

  hints_display_url_panel=true;

Thanks to John Foerch for suggestions.