Remove minibuffer annotation during minibuffer recursion
When the minibuffer is called recursively, the minibuffer annotation
should be hidden or updated appropriately.  For instance, previously
if "C-h c" was pressed while the annotation was shown during a hints
interaction, the annotation would confusingly remain visible while the
minibuffer prompted for a key.  Now the annotation will be hidden
until the user returns to the hints interaction.