reddit-mode: fix handling of current link
There is a sidebar on reddit for recently-viewed items, and items in this
sidebar can sometimes have the 'last-clicked' class, which caused various
commands in reddit-mode to find the wrong link.  This patch makes the
selectors more specific so that only a 'last-clicked' link in the main
content area will be found.