buffer_container: mark hidden buffers inactive
When a buffer is not visible, it can still eats CPU due to animations and
Javascript.  The docShell property 'isActive', introduced in Gecko 2.0,
can be used to mitigate this.  When false, it stops CSS animations, runs
timers less often, and generally causes the inactive buffer to consume a
lot less CPU time.