gmail-mode: remove gmail_focus_primary_frame
According to users of gmail-mode, GMail has been changed in some way that
caused keys to no longer work.  Experimentation showed that
gmail_focus_primary_frame was the source of the problem, in that it was
focusing the wrong frame, but when removed altogether, the page-mode
worked fine.

Tested by Tim Ramsey and Jean-Baptiste Mestelan.