save_uri: update for xulrunner 18
In xulrunner 18, nsITransfer.init has a new parameter, aIsPrivate.  There
is no error or warning to pass a value for this parameter in earlier
versions of xulrunner that we have tried, so we simply always pass it.  If
problems are reported, we can change to a version dispatch like we have
done for some other similar changes.

"aIsPrivate Used to determine the privacy status of the new transfer.  If
true, indicates that the transfer was initiated from a source that desires
privacy."