download-manager: allow download-show to work with paused downloads from a previous process
A new download_info object is created when the user attempts to show
such a download, instead of producing an error in the completer.

Additionally, there is now existing_download_added_hook, which is
called when a download_info object is created in response to a
download from a previous process.