modifiers.s: support for Super key / Windows-logo key
In XULRunner 12 and earlier, Windows-logo key was recognized as the
metaKey modifier.  In 13 and 14, it was not recognized at all.  XULRunner
15 introduced the DOM3 method event.getModifierState, which brought back
support for the Windows-logo key, aka Super key, as the "OS key".  With
this patch, Conkeror maps that to the "s" modifier.

This patch is based on work by Nguyễn Tuấn Anh.

See also:
  https://developer.mozilla.org/en-US/docs/DOM/KeyboardEvent#getModifierState()

Resolves issue443 <http://bugs.conkeror.org/issue443>