load_rc: avoid NS_ERROR_FILE_TARGET_DOES_NOT_EXIST in symlink test
Previously, if there was no rc, the symlink test on the rc path would
throw NS_ERROR_FILE_TARGET_DOES_NOT_EXIST.  This patch introduces a catch
for that error.

file_symlink_p: new util.  takes an nsIFile and returns the value of
file.isSymlink(), but also catches the error and returns false if the file
does not exist.  Note that this cannot be tested with file.exists(),
because that method automatically resolves symlinks.