duckduckgo_anchor_keymap: remove binding of duckduckgo-follow-current (return)
Just use Conkeror's follow-current command.  Calling in to the content js
stopped working with a recent update of duckduckgo.com.

Resolves issue 446.

There may be more work to be done on this because there seem to be some
instances where follow-current does not work.  This fixes it for the main
links, at least.