Debian package: Make alternative dependency only on iceweasel >= 30
Rationale:

Conkeror would work fine with any Iceweasel version greate 3.5. But
any Iceweasel version between 3.5 and 30 pull in the according
Xulrunner package and that won't change retroactively. (It only did
for Iceweasel 30 of which the first package version still built a
Xulrunner package.)

So for Iceweasel/Xulrunner <= 30 it's irrelevant if iceweasel is
installed, too, or not. Conkeror then only needs Xulrunner although it
could use "iceweasel -app" instead if iceweasel is installed.

Thanks to Emilio Pozuelo Monfort for the discussion around that topic.