reddit.js: optimize previous link wrap behavior
Checking the display style in the loop doesn't make much sense because
that would only matter if every link on the page was invisible. Even if
this is the case this loop would still not catch that because
dom_add_class() and reddit_scroll_into_view() would later be called with
prev being null.